feat: Make all classes autoloadable by composer, reorganise files
Although our code was placed in Kohana cascading-filesystem directories
following the Kohana convention, we never actually used (or supported)
Kohana's transparent class replacement.

Move everything to standard PSR-4 autoloading and reorganise files
within the repo to suit.
