+++ rector.php
@@ -8,8 +8,8 @@ use Rector\TypeDeclaration\Rector\ClassMethod\ReturnNeverTypeRector;
 
 return RectorConfig::configure()
     ->withPaths([
-        __DIR__.'/classes',
         __DIR__.'/config',
+        __DIR__.'/src',
         __DIR__.'/tests',
     ])
     ->withRootFiles()
@@ -22,11 +22,11 @@ return RectorConfig::configure()
     ->withSkip([
         ClosureToArrowFunctionRector::class => [
             // Needs to be a traditional function in order to restrict the scope
-            __DIR__.'/classes/Ingenerator/KohanaView/Renderer/HTMLRenderer.php',
+            __DIR__.'/src/Renderer/HTMLRenderer.php',
         ],
         ReturnNeverTypeRector::class => [
             // This stub function shouldn't return `never` as it will actually "return" a string
-            __DIR__.'/classes/raw_fn_stub.php',
+            __DIR__.'/src/raw_fn_stub.php',
         ],
     ])
     ->withImportNames(
