+++ src/Exception/InvalidTemplateContentException.php
