+++ src/Exception/InvalidViewVarAssignmentException.php
