+++ src/TemplateManager.php
