+++ src/ViewModel/PageLayout/StaticPageContentView.php
