+++ src/raw_fn_stub.php
