+++ tests/bootstrap.php
@@ -1,14 +1,6 @@
 <?php
 
-use Composer\Autoload\ClassLoader;
-
 // Bootstrap for running unit tests
 define('TEST_ROOT_PATH', realpath(__DIR__).'/');
 require_once __DIR__.'/../koharness_bootstrap.php';
 require_once KOHARNESS_SRC.'helper_classes/Session/Fake.php';
-
-// Autoload mocks and test-support helpers that should not autoload in the main app
-$mock_loader = new ClassLoader();
-$mock_loader->addPsr4('test\\mock\\', [__DIR__.'/mock/']);
-$mock_loader->addPsr4('test\\unit\\', [__DIR__.'/unit/']);
-$mock_loader->register();
