+++ tests/unit/TemplateCompilerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace test\unit\Ingenerator\KohanaView;
+namespace test\unit;
 
 use Ingenerator\KohanaView\Exception\InvalidTemplateContentException;
 use Ingenerator\KohanaView\TemplateCompiler;
